Fix change project structure for QT migration + test ploting threads
