+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = []
 package-mode = false
 
 [tool.poetry.dependencies]
-python = "^3.12"
+python = ">=3.12,<3.14"
 annotated-types = "0.7.0"
 appdirs = "1.4.4"
 asteval = "1.0.5"
@@ -95,6 +95,7 @@ pytest = "8.3.3"
 pylint = "^3.3.1"
 black = "^24.10.0"
 flake8 = "^7.1.1"
+pyside6 = "^6.8.0.2"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
